+++ typescript/cli/src/tests/ethereum/warp/warp-apply.e2e-test.ts
@@ -20,17 +20,17 @@ import {
 } from '@hyperlane-xyz/utils';
 
 import { readYamlOrJson, writeYamlOrJson } from '../../../utils/files.js';
+import { deployOrUseExistingCore } from '../commands/core.js';
 import {
-  deployOrUseExistingCore,
   exportWarpConfigsToFilePaths,
   getDeployedWarpAddress,
-  updateOwner,
 } from '../commands/helpers.js';
 import {
   hyperlaneWarpApply,
   hyperlaneWarpApplyRaw,
   hyperlaneWarpDeploy,
   readWarpConfig,
+  updateOwner,
 } from '../commands/warp.js';
 import {
   ANVIL_DEPLOYER_ADDRESS,
