+++ typescript/cli/src/tests/ethereum/warp/warp-bridge-utils.ts
@@ -14,13 +14,13 @@ import {
 import { Address } from '@hyperlane-xyz/utils';
 
 import { readYamlOrJson, writeYamlOrJson } from '../../../utils/files.js';
+import { deployOrUseExistingCore } from '../commands/core.js';
+import { deploy4626Vault, deployToken } from '../commands/helpers.js';
 import {
-  deploy4626Vault,
-  deployOrUseExistingCore,
-  deployToken,
+  generateWarpConfigs,
+  hyperlaneWarpDeploy,
   sendWarpRouteMessageRoundTrip,
-} from '../commands/helpers.js';
-import { generateWarpConfigs, hyperlaneWarpDeploy } from '../commands/warp.js';
+} from '../commands/warp.js';
 import {
   ANVIL_KEY,
   CHAIN_2_METADATA_PATH,
