+++ typescript/cli/src/tests/ethereum/warp/warp-check.e2e-test.ts
@@ -35,7 +35,8 @@ import {
 } from '@hyperlane-xyz/utils';
 
 import { readYamlOrJson, writeYamlOrJson } from '../../../utils/files.js';
-import { deployOrUseExistingCore, deployToken } from '../commands/helpers.js';
+import { deployOrUseExistingCore } from '../commands/core.js';
+import { deployToken } from '../commands/helpers.js';
 import {
   hyperlaneWarpApply,
   hyperlaneWarpCheckRaw,
