+++ typescript/cli/src/tests/ethereum/warp/warp-extend-config.e2e-test.ts
@@ -12,12 +12,10 @@ import {
 import { addressToBytes32 } from '@hyperlane-xyz/utils';
 
 import { readYamlOrJson, writeYamlOrJson } from '../../../utils/files.js';
+import { deployOrUseExistingCore } from '../commands/core.js';
+import { getDomainId } from '../commands/helpers.js';
 import {
-  deployOrUseExistingCore,
   extendWarpConfig,
-  getDomainId,
-} from '../commands/helpers.js';
-import {
   hyperlaneWarpApply,
   hyperlaneWarpDeploy,
   readWarpConfig,
